feat: Fetch external names on render
