+++ src/components/WorldListPage_WorldsForEnsOwnersFeature/WorldListPage.container.ts
@@ -2,8 +2,8 @@ import { connect } from 'react-redux'
 import { push } from 'connected-react-router'
 import { isLoadingType } from 'decentraland-dapps/dist/modules/loading/selectors'
 import { RootState } from 'modules/common/types'
-import { FETCH_ENS_LIST_REQUEST } from 'modules/ens/actions'
-import { getENSByWallet, getError as getENSError, getLoading } from 'modules/ens/selectors'
+import { FETCH_ENS_LIST_REQUEST, FETCH_EXTERNAL_NAMES_REQUEST, fetchExternalNamesRequest } from 'modules/ens/actions'
+import { getENSByWallet, getError as getENSError, getLoading as getLoadingENS } from 'modules/ens/selectors'
 import { FETCH_WORLD_DEPLOYMENTS_REQUEST } from 'modules/deployment/actions'
 import { getDeploymentsByWorlds, getError as getDeploymentsError, getLoading as getDeploymentsLoading } from 'modules/deployment/selectors'
 import { FETCH_LANDS_REQUEST } from 'modules/land/actions'
@@ -21,10 +21,11 @@ const mapState = (state: RootState): MapStateProps => ({
   projects: getProjects(state),
   error: getENSError(state)?.message ?? getDeploymentsError(state) ?? undefined,
   isLoading:
-    isLoadingType(getLoading(state), FETCH_ENS_LIST_REQUEST) ||
+    isLoadingType(getLoadingENS(state), FETCH_ENS_LIST_REQUEST) ||
     isLoadingType(getDeploymentsLoading(state), FETCH_WORLD_DEPLOYMENTS_REQUEST) ||
     isLoadingType(getLandsLoading(state), FETCH_LANDS_REQUEST) ||
     isLoadingType(getLoadingWorlds(state), FETCH_WORLD_DEPLOYMENTS_REQUEST) ||
+    isLoadingType(getLoadingENS(state), FETCH_EXTERNAL_NAMES_REQUEST) ||
     isLoggingIn(state),
   isLoggedIn: isLoggedIn(state),
   worldsWalletStats: getConnectedWalletStats(state)
@@ -32,7 +33,8 @@ const mapState = (state: RootState): MapStateProps => ({
 
 const mapDispatch = (dispatch: MapDispatch): MapDispatchProps => ({
   onNavigate: path => dispatch(push(path)),
-  onFetchWorldsWalletStats: () => dispatch(fetchWorldsWalletStatsRequest())
+  onFetchWorldsWalletStats: () => dispatch(fetchWorldsWalletStatsRequest()),
+  onFetchExternalNames: () => dispatch(fetchExternalNamesRequest())
 })
 
 export default connect(mapState, mapDispatch)(WorldListPage)
